Move test to existing module
