+++ tests/test_options.py
@@ -180,6 +180,24 @@ def test_nargs_envvar(runner):
     assert result.output == 'x|1\ny|2\n'
 
 
+def test_custom_validation_unicode(runner):
+    def validate_pos_int(ctx, value):
+        if value < 0:
+            raise click.BadParameter(u'That’s no good!')
+        return value
+
+    @click.command()
+    @click.option('--foo', callback=validate_pos_int, default=1)
+    def cmd(foo):
+        click.echo(foo)
+
+    result = runner.invoke(cmd, ['--foo', '-1'])
+    assert 'Invalid value for "--foo": That’s no good' in result.output
+
+    result = runner.invoke(cmd, ['--foo', '42'])
+    assert result.output == '42\n'
+
+
 def test_custom_validation(runner):
     def validate_pos_int(ctx, value):
         if value < 0:
